small optimizations: vertex index size_t -> uint, i++ -> ++i

+++ include/BaseGraph/types.h
@@ -9,7 +9,7 @@
 namespace BaseGraph {
 
 /// Integer that identifies a vertex.
-typedef size_t VertexIndex;
+typedef unsigned int VertexIndex;
 
 /// Empty type used for unlabeled graphs.
 struct NoLabel {
